+++ src/api.rs
@@ -24,7 +24,7 @@ use std::fmt;
 
 use crate::error::Error;
 use crate::error::Error::{ClientErrorResponse, NotFound, ServerErrorResponse};
-use crate::requests::EmptyPayload;
+use crate::requests::{EmptyPayload, StreamParams};
 use crate::responses::{
     DeprecatedFeatureList, FeatureFlag, FeatureFlagList, FeatureFlagStability, FeatureFlagState,
     GetMessage, OAuthConfiguration, SchemaDefinitionSyncStatus, WarmStandbyReplicationStatus,
@@ -618,6 +618,15 @@ where
         Ok(response)
     }
 
+    /// Returns information about a stream.
+    pub async fn get_stream_info(
+        &self,
+        virtual_host: &str,
+        name: &str,
+    ) -> Result<responses::QueueInfo> {
+        self.get_queue_info(virtual_host, name).await
+    }
+
     /// Returns information about an exchange.
     pub async fn get_exchange_info(
         &self,
@@ -681,6 +690,27 @@ where
     pub async fn declare_queue(&self, vhost: &str, params: &QueueParams<'_>) -> Result<()> {
         let _response = self
             .http_put(path!("queues", vhost, params.name), params, None, None)
+            .await?;
+        Ok(())
+    }
+
+    pub async fn declare_stream(&self, vhost: &str, params: &StreamParams<'_>) -> Result<()> {
+        let mut m: Map<String, Value> = Map::new();
+
+        if let Some(m2) = params.arguments.clone() {
+            m.extend(m2);
+        };
+
+        if let Some(val) = params.max_length_bytes {
+            m.insert("max_length_bytes".to_owned(), json!(val));
+        };
+        if let Some(val) = params.max_segment_length_bytes {
+            m.insert("max_segment_length_bytes".to_owned(), json!(val));
+        };
+
+        let q_params = QueueParams::new_stream(params.name, Some(m));
+        let _response = self
+            .http_put(path!("queues", vhost, params.name), &q_params, None, None)
             .await?;
         Ok(())
     }
@@ -805,6 +835,10 @@ where
             .http_delete(path!("queues", vhost, name), excludes, None)
             .await?;
         Ok(())
+    }
+
+    pub async fn delete_stream(&self, vhost: &str, name: &str, idempotently: bool) -> Result<()> {
+        self.delete_queue(vhost, name, idempotently).await
     }
 
     pub async fn delete_exchange(&self, vhost: &str, name: &str, idempotently: bool) -> Result<()> {