+++ tests/async_queue_tests.rs
@@ -67,7 +67,7 @@ async fn test_async_declare_a_quorum_queue() {
 }
 
 #[tokio::test]
-async fn test_async_declare_a_stream() {
+async fn test_async_declare_a_stream_with_declare_queue() {
     let endpoint = endpoint();
     let rc = Client::new(&endpoint, USERNAME, PASSWORD);
     let vhost = "/";